% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lftp_settings.R
\name{check_lftp_settings}
\alias{check_lftp_settings}
\title{Check `lftp` settings availability in session options}
\usage{
check_lftp_settings(verbose = FALSE)
}
\arguments{
\item{verbose}{Boolean. Whether or not to display a short message after the check}
}
\value{
Nothing. Throws an error in case of missing functions
}
\description{
This function ensures that the `lftp` settings are accessible for the package's functions or returns an error
}
\seealso{
[create_lftp_settings()] and [parse_lftp_settings()]
}
