% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_accession_files_from_ftp.R
\name{download_accession_files_from_ftp}
\alias{download_accession_files_from_ftp}
\title{Download Files from the GWAS Catalog FTP Server}
\usage{
download_accession_files_from_ftp(
  accession_file_links,
  download_directory = NA,
  create_accession_directory = FALSE,
  overwrite_existing_files = FALSE,
  lftp_settings = NA
)
}
\arguments{
\item{accession_file_links}{A named list with a vector of character strings.
Use the output of `get_harmonised_accession_file_links()` or
`get_accession_file_links()` as the input for this function.}

\item{download_directory}{character string. Output file directory. Downloaded
files will retain their names.}

\item{create_accession_directory}{logical. `TRUE`/`FALSE` for whether a new
directory named after `study_accession` should be created in
`download_directory`  to save all downloaded files in.}

\item{overwrite_existing_files}{logical. `TRUE`/`FALSE` for whether identical
output files found in `download_directory` should be overwritten.}

\item{lftp_settings}{named list. Use `create_lftp_settings()` to create.}
}
\value{
Nothing. Will download files into `download_directory`.
}
\description{
This files takes a named list of FTP links as input and downloads all files to a local directory. To download all data for a GWAS Catalog study, most users will probably want to use `download_all_accession_data()` instead of this function.
}
\seealso{
[download_all_accession_data()]
}
