% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_accession_data.R
\name{download_all_accession_data}
\alias{download_all_accession_data}
\title{Download all GWAS Catalog Data for a Study Accession}
\usage{
download_all_accession_data(
  study_accession,
  harmonised_list = NA,
  directory_list = NA,
  lftp_settings = NA,
  download_directory = NA,
  create_accession_directory = FALSE,
  overwrite_existing_files = FALSE,
  return_meta_data = TRUE
)
}
\arguments{
\item{study_accession}{single character string Input. GWAS Catalog study ID
to download data for. Must be an accession number beginning with “GCST”.
Use the package `{gwasrapidd}` or browse the GWAS Catalog website
(https://www.ebi.ac.uk/gwas/) to get accessions for studies you want to
download data from.}

\item{harmonised_list}{character vector. Use `get_harmonised_list()` to
create. Contains the content of the FTP file
`ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/harmonised_list.txt`.
Should be downloaded once and saved in a variable to be then passed to
functions using this field.}

\item{directory_list}{character vector. Use `get_directory_list()` to create.
Should be downloaded once and saved in a variable to be then passed to
functions using this field.}

\item{lftp_settings}{named list. Use `create_lftp_settings()` to create.}

\item{download_directory}{character string. Output file directory. Downloaded
files will retain their names.}

\item{create_accession_directory}{logical. `TRUE`/`FALSE` for whether a new
directory named after `study_accession` should be created in
`download_directory`  to save all downloaded files in.}

\item{overwrite_existing_files}{logical. `TRUE`/`FALSE` for whether identical
output files found in `download_directory` should be overwritten.}

\item{return_meta_data}{logical. `TRUE`/`FALSE` for whether to return the
study accessions meta-data (if available as `yaml` file) in a data.table
with one row per found file.}
}
\value{
If `return_meta_data` is set to `TRUE`, returns a data.table
  containing one row per meta-data yaml file found for the given study.
}
\description{
`download_all_accession_data()` uses the `lftp` FTP client to look up and
download all available data (meta-data, summary statistics, etc.) for a
study, identified by the accession `GCST[...]`, from the GWAS Catalog. For
instance, to download all data of the study listed at
`https://www.ebi.ac.uk/gwas/studies/GCST90162626`, use `GCST90162626` as the
input for `study_accession`.
}
\seealso{
[create_lftp_settings], [get_harmonised_list] and
  [get_directory_list] that are used to create required input files for this
  function.
}
