% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_multiple_accession_meta_data.R
\name{download_multiple_accession_meta_data}
\alias{download_multiple_accession_meta_data}
\title{Look up, download and parse meta-data for several study accessions}
\usage{
download_multiple_accession_meta_data(
  study_accessions,
  harmonised_list = NA,
  directory_list = NA,
  lftp_settings = NA
)
}
\arguments{
\item{study_accessions}{Character vector. GWAS Catalog study ID(s) to download
meta-data for. Must be an accession number beginning with “GCST”. Use the
package `\{gwasrapidd\}` or browse the GWAS Catalog website
(https://www.ebi.ac.uk/gwas/) to get accessions for studies you want to
download data from.}

\item{harmonised_list}{character vector. Use `get_harmonised_list()` to
create. Contains the content of the FTP file
`ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/harmonised_list.txt`.
Should be downloaded once and saved in a variable to be then passed to
functions using this field.}

\item{directory_list}{character vector. Use `get_directory_list()` to create.
Should be downloaded once and saved in a variable to be then passed to
functions using this field.}

\item{lftp_settings}{named list. Use `create_lftp_settings()` to create.}
}
\value{
a data.table with n=`length(study_accessions)` rows. Studies with no
  available meta-data will have `NA` entries for all fields.
}
\description{
Given a vector of study accessions (GWAS Catalog study IDs starting with
`GCST[...]`), download all available meta data and parse them into a
data.table with one row per available meta-data `yaml` file.
}
