% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bucket.R
\name{find_bucket}
\alias{find_bucket}
\title{Find the FTP directory of a study accession}
\usage{
find_bucket(study_accession, directory_list = NA)
}
\arguments{
\item{study_accession}{single character string Input. GWAS Catalog study ID
to download data for. Must be an accession number beginning with “GCST”.
Use the package `\{gwasrapidd\}` or browse the GWAS Catalog website
(https://www.ebi.ac.uk/gwas/) to get accessions for studies you want to
download data from.}

\item{directory_list}{character vector. Use `get_directory_list()` to create.
Should be downloaded once and saved in a variable to be then passed to
functions using this field.}
}
\value{
The relative path from
  `ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/` of the directory
  containing the data of the input study accession (e.g.
  `	GCST000001-GCST001000/`).
}
\description{
For a given GWAS Catalog study accession, identify the directory containing
study data on the GWAS Catalog FTP server at
`https://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/`.
}
\details{
On the GWAS Catalog FTP server, all study directories are located in
subdirectories covering a range of study IDs. For example, the directory
`GCST000001-GCST001000/` contains all studies whose numeric part of their ID
lies between 1-1000. This function simply returns this subdirectory.
}
