% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accession_file_links.R
\name{get_accession_file_links}
\alias{get_accession_file_links}
\title{Find FTP links to all non-harmonised files for a GWAS Catalog study accession}
\usage{
get_accession_file_links(
  study_accession,
  directory_list = NA,
  lftp_settings = NA
)
}
\arguments{
\item{study_accession}{single character string Input. GWAS Catalog study ID
to download data for. Must be an accession number beginning with “GCST”.
Use the package `\{gwasrapidd\}` or browse the GWAS Catalog website
(https://www.ebi.ac.uk/gwas/) to get accessions for studies you want to
download data from.}

\item{directory_list}{character vector. Use `get_directory_list()` to create.
Should be downloaded once and saved in a variable to be then passed to
functions using this field.}

\item{lftp_settings}{named list. Use `create_lftp_settings()` to create.}
}
\value{
A named list containing a character string vector with the FTP links
  to all non-harmonised files for the input study accession.
}
\description{
This function looks up and returns a vector containing FTP links to all
non-harmonised files for a given GWAS Catalog study accession.  To get links
to harmonised files where available, use the function
[get_harmonised_accession_file_links()].
}
