% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accession_meta_data.R
\name{get_accession_meta_data}
\alias{get_accession_meta_data}
\title{From a list of accession file links, download the raw meta-data}
\usage{
get_accession_meta_data(accession_file_links, lftp_settings = NA)
}
\arguments{
\item{accession_file_links}{A named list with a vector of character strings.
Use the output of `get_harmonised_accession_file_links()` or
`get_accession_file_links()` as the input for this function.}

\item{lftp_settings}{named list. Use `create_lftp_settings()` to create.}
}
\value{
A named list of raw and unparsed character strings containing the
  content of the meta-data yaml file
}
\description{
This function tries to download the meta-data for a GWAS Catalog study
accession that is contained in a `yaml` file (ending on `-meta.yaml`) where
available. This function receives a names list as input that is the output of
the functions `get_harmonised_accession_file_links()` or
`get_accession_file_links()`.
}
\seealso{
[parse_raw_meta_data()]
}
