% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_directory_list.R
\name{get_directory_list}
\alias{get_directory_list}
\title{Download a list of all sub-directories of the GWAS Catalog FTP directory
`summary_statistics/`}
\usage{
get_directory_list(lftp_settings = NA)
}
\arguments{
\item{lftp_settings}{named list. Use `create_lftp_settings()` to create.}
}
\value{
A character vector containing all directories of
  `https://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/`.
}
\description{
Function to get a list of all direct subdirectories of the GWAS Catalog FTP
directory `https://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/`. The
list is used in several functions of this package and should therefore be
downloaded once to be used as input for those functions instead of
downloading it with every function call.
}
