% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_harmonised_list.R
\name{get_harmonised_list}
\alias{get_harmonised_list}
\title{Save the list of all harmonised files from the GWAS Catalog}
\usage{
get_harmonised_list()
}
\value{
A character vector containing the content of
  `https://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/harmonised_list.txt`
}
\description{
The GWAS Catalog FTP server contains a file with all harmonised data sets at
`https://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/harmonised_list.txt`.
This list is used by several function of this package. With this function,
the list is downloaded and can be saved in a variable to avoid repeated
remote access of this file. See `https://www.ebi.ac.uk/gwas/docs/faq` for
more information on harmonised data in the GWAS Catalog.
}
