% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwascatftp-package.R
\docType{package}
\name{gwascatftp-package}
\alias{gwascatftp}
\alias{gwascatftp-package}
\title{gwascatftp: GWAS Catalog FTP Interface}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Carl Beuchel \email{carl.beuchel@charite.de} (\href{https://orcid.org/0000-0003-3224-3894}{ORCID})

}
\keyword{internal}
