% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_lftp_settings.R
\name{parse_lftp_settings}
\alias{parse_lftp_settings}
\title{Make `lftp` settings available}
\usage{
parse_lftp_settings(lftp_settings = NA)
}
\arguments{
\item{lftp_settings}{named list. Use `create_lftp_settings()` to create.}
}
\value{
Local variables named after the elements of  containing the
}
\description{
This function exports the values in `lftp_settings` to local variables in the
function's parent environment.
}
\seealso{
[create_lftp_settings()]
}
