% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_lftp.R
\name{install_lftp}
\alias{install_lftp}
\title{Provide information for installing `lftp`}
\usage{
install_lftp()
}
\value{
If `lftp` is already installed and can be found via `whereis` system command, returns the full path to the binary. If not, it prints a helpful message about where and how to install `lftp` from.
}
\description{
This package depends on the command-line program `lftp` that must be
available on the system to function. It will have to be installed manually by
the package user, either globally, or locally using a local environment
manager such as `conda`.
}
