% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_lftp_settings.R
\name{parse_lftp_settings}
\alias{parse_lftp_settings}
\title{Make `lftp` settings available}
\usage{
parse_lftp_settings(
  lftp_settings = create_lftp_settings(),
  check_settings = FALSE
)
}
\arguments{
\item{lftp_settings}{List. The settings to parse as session-wide options. By default
this list is automatically created by calling `create_lftp_settings()` but when settings
need to be manually adapted, the user needs to supply the custom settings
 to `create_lftp_settings()`.}

\item{check_settings}{Logical. Whether or not to run `check_lftp_settings()` 
on the parsed settings. Default is FALSE}
}
\value{
Nothing. Check `options()` to see changes.
}
\description{
This function exports the values in `lftp_settings` to the session's options in the
function's parent environment. This overrides all settings, so make sure the 
list supplied to `lftp_settings` contains all neccessary elements!
}
\seealso{
[create_lftp_settings()]
}
